testcontainer refactoring
Based on observation, localstack containers weren't being shutdown
properly.

It turns out static bindings don't get shut down, so we need
https://github.com/testcontainers/testcontainers-rs/issues/707#issuecomment-2309718920
unti https://github.com/testcontainers/testcontainers-rs/issues/577.
