+++ server/Cargo.toml
@@ -24,7 +24,7 @@ reqwest = { version = "0.12", default-features = false, features = [
 axum = "0.7"
 serde = "1.0"
 serde_json = "1.0"
-tokio = { version = "1", features = ["macros"] }
+tokio = { version = "1", features = ["macros", "sync"] }
 image = { version = "0.25", features = ["rayon"] }
 
 miniaturs_shared = { path = "../shared" }
@@ -36,5 +36,6 @@ bytes = "1.7"
 tower-http = { version = "0.6.1", features = ["catch-panic"] }
 
 [dev-dependencies]
+ctor = "0.2.8"
 testcontainers = { version = "0.23" }
 testcontainers-modules = { version = "0.11", features = ["localstack"] }
